+++ gulpfile.js/tasks/clean.js
@@ -1,3 +1,5 @@
+'use strict'
+
 var del = require('del')
 var gulp = require('gulp')
 var config = require('../config')